Upgrade @vitejs/plugin-react to version 4.2.1
The @vitejs/plugin-react package has been upgraded from version 3.1.0 to version 4.2.1. This change updates packages in both package.json and package-lock.json. Several dependencies associated with the plugin are also updated or added as part of this change, enhancing the plugin's functionality and compatibility with recent versions of other software.